Add files via upload
A credit score is a three-digit number that rates your creditworthiness. FICO scores range from 300 to 850. The higher the score, the more likely you are to get approved for loans and for better rates. A credit score is a number that depicts a consumer’s creditworthiness. FICO scores range from 300 to 850